requirements.txt
numpy~=1.19.5
nibabel~=3.2.1
vtk==8.1.2
mayavi==4.6.2
matplotlib~=3.4.2
scikit-learn~=1.0.1
pysurfer~=0.11.0
mne>=0.20,<0.21